+++ libs/drawer/src/lib/drawer-container/drawer-container-proxy.component.ts
@@ -1,0 +1,14 @@
+import { Component, ViewChild } from '@angular/core';
+
+import { DrawerContainerComponent } from './drawer-container.component';
+
+@Component({
+  selector: 'spy-drawer-container-proxy',
+  template: `
+    <spy-drawer-container></spy-drawer-container>
+  `,
+})
+export class DrawerContainerProxyComponent {
+  @ViewChild(DrawerContainerComponent, { static: true })
+  container!: DrawerContainerComponent;
+}