+++ libs/drawer/src/lib/drawer-interception.ts
@@ -1,0 +1,7 @@
+import { InterceptionEvent } from '@spryker/interception';
+
+export class DrawerCloseInterceptionEvent extends InterceptionEvent {}
+
+export class DrawerMaximizeInterceptionEvent extends InterceptionEvent {}
+
+export class DrawerMinimizeInterceptionEvent extends InterceptionEvent {}