+++ libs/locale/locales/en/src/index.ts
@@ -2,7 +2,7 @@ import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { LocaleRecord, provideLocaleRecordsFactory } from '@spryker/locale';
 
-// Locales version update 1
+// Locales version update 2
 
 export const EN_LOCALE = 'en';