+++ libs/modal/src/lib/strategies/confirm/confirm.module.ts
@@ -1,0 +1,13 @@
+import { CommonModule } from '@angular/common';
+import { NgModule } from '@angular/core';
+import { ButtonModule } from '@spryker/button';
+import { OfTypePipeModule } from '@spryker/utils';
+
+import { ConfirmModalComponent } from './confirm.component';
+
+@NgModule({
+  imports: [CommonModule, ButtonModule, OfTypePipeModule],
+  declarations: [ConfirmModalComponent],
+  entryComponents: [ConfirmModalComponent],
+})
+export class ConfirmModalModule {}