+++ libs/modal/src/lib/strategies/index.ts
@@ -1,0 +1,4 @@
+export * from './component.strategy';
+export * from './template.strategy';
+export * from './confirm';
+export * from './html.strategy';