+++ libs/unsaved-changes.guard.drawer/src/i18n/en.ts
@@ -1,0 +1,10 @@
+import { I18nLocaleDataPackage } from '@spryker/locale';
+
+export default {
+  name: 'unsaved-changes.guard.drawer',
+  data: {
+    title: 'Discard unsaved changes?',
+    ok: 'Discard',
+    cancel: 'Cancel',
+  },
+} as I18nLocaleDataPackage;